requirements.txt
streamlit
opencv
opencv-python-headless
numpy
tensorflow
Pillow
